requirements.txt
dj-database-url==0.5.0
Django==2.0.2
django-bootstrap3==9.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
MaterialDjango==0.2.5
psycopg2==2.7.4
pytz==2018.3
whitenoise==3.3.1
